change default layout to be same as root, remove root layout
