--- components/layouts/Root.js
@@ -1,97 +1,0 @@
-import React, { useState } from 'react';
-import { ThemeProvider } from 'styled-components';
-import { DefaultSeo, LogoJsonLd, SocialProfileJsonLd } from 'next-seo';
-import { useRouter } from 'next/router';
-// import LogRocket from 'logrocket';
-import * as Sentry from '@sentry/browser';
-import { isNil, map } from 'lodash';
-
-import GlobalStyle from '../../styles/globalStyle';
-import baseTheme from '../../styles/baseTheme';
-import Meta from '../Meta';
-import Header from '../Header/root';
-import Footer from '../Footer';
-import { rootSeo, socialConstants } from '../../utilities';
-import User from '../User';
-import { useFetchUser } from '../../hooks/user';
-import {
-  StyledPage,
-  CorePage,
-  PageDiv,
-  InnerPage,
-} from '../shared/StandardStyles';
-
-const Default = ({ children }) => {
-  const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
-
-  const { user, loading } = useFetchUser();
-  const router = useRouter();
-
-  if (!isNil(user)) {
-    document.tidioIdentify = {
-      distinct_id: user.id,
-      email: user.email,
-    };
-
-    // LogRocket.identify(user.id, {
-    //   email: user.email,
-    // });
-
-    Sentry.configureScope(scope => {
-      scope.setUser({
-        email: user.email,
-        id: user.id,
-      });
-      // scope.setExtra('sessionURL', LogRocket.sessionURL);
-    });
-  }
-
-  const socialUrls = map(
-    socialConstants.thatSocialLinks,
-    socialLink => socialLink.url,
-  );
-
-  return (
-    <ThemeProvider theme={baseTheme}>
-      <>
-        <GlobalStyle />
-        <StyledPage>
-          <Meta />
-          <DefaultSeo
-            {...rootSeo}
-            canonical={`https://www.thatconference.com/${router.pathname}`}
-          />
-          <LogoJsonLd
-            logo="https://www.thatconference.com/svgs/THAT.svg"
-            url="https://www.thatconference.com"
-          />
-          <SocialProfileJsonLd
-            type="Organization"
-            name="THAT"
-            url="http://www.thatconference.com"
-            logo="https://www.thatconference.com/svgs/THAT.svg"
-            sameAs={socialUrls}
-          />
-          <CorePage>
-            <User user={user} loading={loading}>
-              <PageDiv mobileMenuOpen={mobileMenuOpen}>
-                <Header
-                  user={user}
-                  loading={loading}
-                  mobileMenuOpen={mobileMenuOpen}
-                  setMobileMenuOpen={setMobileMenuOpen}
-                />
-                <InnerPage>
-                  {React.cloneElement(children, { user, loading })}
-                </InnerPage>
-                <Footer modifiers="site" />
-              </PageDiv>
-            </User>
-          </CorePage>
-        </StyledPage>
-      </>
-    </ThemeProvider>
-  );
-};
-
-export default Default;
