+++ pages/tx/2021/index.js
@@ -3,13 +3,11 @@
  */
 
 import React from 'react';
-import RootLayout from '../../../components/layouts/Root';
+
 import ComingSoon from '../../../components/EventPages/ComingSoon';
 
 const Index = () => {
   return <ComingSoon />;
 };
-
-Index.Layout = RootLayout;
 
 export default Index;
